apps/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-jose[cryptography]
passlib
python-multipart
PyPDF2
python-docx
